+++ examples/plot_power.py
